Correct JSON issues
Placed a few new fixes in place to correct for json problems that have been happening repeatedly.

We are making changes to the array as we are building them. We are then addressing the strange issues resulting from that when it reloads with the slice to remove the double quotes.

This is currently working in test on my stream.